# Create Composite Index for `xform_id` and `id` fields